api and gate
