+++ game-api/src/main/java/org/game/api/db/MongoDbService.java
@@ -1,0 +1,23 @@
+package org.game.api.db;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.mongodb.MongoDbFactory;
+import org.springframework.stereotype.Component;
+
+/**
+ * mongodb服务管理
+ * 
+ * @author JiangZhiYong
+ * @mail 359135103@qq.com
+ */
+@Component
+public class MongoDbService {
+
+	protected final MongoDbFactory dbFactory;
+
+	@Autowired
+	public MongoDbService(MongoDbFactory dbFactory) {
+		this.dbFactory = dbFactory;
+	}
+
+}
