+++ game-api/src/main/java/org/game/api/script/IAccountScript.java
@@ -1,0 +1,20 @@
+package org.game.api.script;
+
+import com.jzy.javalib.base.script.IScript;
+import io.grpc.stub.StreamObserver;
+import org.jzy.game.proto.LoginRequest;
+import org.jzy.game.proto.LoginResponse;
+
+/**
+ * 登录
+ * @author jzy
+ */
+public interface IAccountScript extends IScript {
+
+    /**
+     * 登录|注册
+     * @param request
+     * @param responseObserver
+     */
+    void login(LoginRequest request, StreamObserver<LoginResponse> responseObserver);
+}
