+++ game-api/src/main/java/org/game/api/script/IBillingScript.java
@@ -1,0 +1,11 @@
+package org.game.api.script;
+
+
+import com.jzy.javalib.base.script.IScript;
+
+/**
+ * 充值
+ * @author jzy
+ */
+public interface IBillingScript extends IScript {
+}
