+++ game-api/src/main/java/org/game/api/service/ApiExecutorService.java
@@ -1,15 +1,16 @@
-package org.jzy.game.gate.service;
+package org.game.api.service;
 
 import com.jzy.javalib.network.scene.AbstractExecutorService;
 import org.springframework.stereotype.Service;
 
+
 /**
- * 线程执行
- *
+ * 线程管理
  * @author JiangZhiYong
  * @mail 359135103@qq.com
  */
 @Service
-public class GateExecutorService extends AbstractExecutorService {
+public class ApiExecutorService extends AbstractExecutorService {
+
 
 }
