+++ game-api/src/main/java/org/game/api/service/BillingService.java
@@ -1,0 +1,11 @@
+package org.game.api.service;
+
+import org.springframework.stereotype.Service;
+
+/**
+ * 支付相关
+ * @author jzy
+ */
+@Service
+public class BillingService {
+}
