--- game-api/src/main/java/org/mmo/login/script/IBillingScript.java
@@ -1,10 +1,0 @@
-package org.mmo.login.script;
-
-import org.mmo.engine.script.IScript;
-
-/**
- * 充值
- * @author jzy
- */
-public interface IBillingScript extends IScript {
-}
