--- game-api/src/main/java/org/mmo/login/service/BillingService.java
@@ -1,11 +1,0 @@
-package org.mmo.login.service;
-
-import org.springframework.stereotype.Service;
-
-/**
- * 支付相关
- * @author jzy
- */
-@Service
-public class BillingService {
-}
