--- game-api/src/main/java/org/mmo/login/service/LoginExecutorService.java
@@ -1,16 +1,0 @@
-package org.mmo.login.service;
-
-import org.mmo.engine.thread.AbstractExecutorService;
-import org.springframework.stereotype.Service;
-
-
-/**
- * 线程管理
- * @author JiangZhiYong
- * @mail 359135103@qq.com
- */
-@Service
-public class LoginExecutorService extends AbstractExecutorService {
-
-
-}
