+++ game-api/src/main/resources/application.properties
@@ -5,13 +5,13 @@ logging.level.org.mmo=DEBUG
 logging.pattern.console= "%d{yyyy-MM-dd HH:mm:ss}[%15t] %F:%L - [ %p ] %m%n"
 # Logging pattern for file
 logging.pattern.file= "%d{yyyy-MM-dd HH:mm:ss} [%thread] %-5level %logger{36} - %msg%n"
-logging.file.name=../log/login.log
+logging.file.name=../log/api.log
 
 
-#login
-login.id=@login.id@
-login.privateIp=@login.privateIp@
-login.rpcPort=@login.rpcPort@
+#api
+api.id=@api.id@
+api.privateIp=@api.privateIp@
+api.rpcPort=@api.rpcPort@
 
 
 # global
