+++ game-api/src/main/resources/filter/api_jzy.properties
@@ -1,7 +1,7 @@
 #api
-api.id=2
+api.id=1
 api.privateIp=192.168.110.2
-api.rpcPort=7001
+api.rpcPort=7000
 
 global.profile=jzy
 global.zookeeperUrl=192.168.110.2:2181
