--- game-api/src/main/resources/filter/login_jzy.properties
@@ -1,20 +1,0 @@
-#login
-login.id=1
-login.privateIp=192.168.110.2
-login.rpcPort=7000
-
-global.profile=jzy
-global.zookeeperUrl=192.168.110.2:2181
-
-
-#mongodb 游戏数据
-spring.data.mongodb.host=192.168.110.2
-spring.data.mongodb.port=27017
-spring.data.mongodb.authentication-database=admin
-spring.data.mongodb.database=mmo
-spring.data.mongodb.password=123456
-spring.data.mongodb.username=admin
-
-#mongodb配置数据
-config.mongodb.url=mongodb://admin:123456@192.168.110.2:27017
-config.mongodb.database=mmo_config
