+++ game-message/pom.xml
@@ -14,9 +14,24 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.jzy.javalib</groupId>
-            <artifactId>javalib-network</artifactId>
-            <version>0.0.4</version>
+            <groupId>com.google.protobuf</groupId>
+            <artifactId>protobuf-java</artifactId>
+            <version>3.19.6</version>
+        </dependency>
+        <dependency>
+            <groupId>io.grpc</groupId>
+            <artifactId>grpc-netty-shaded</artifactId>
+            <version>1.45.0</version>
+        </dependency>
+        <dependency>
+            <groupId>io.grpc</groupId>
+            <artifactId>grpc-protobuf</artifactId>
+            <version>1.45.0</version>
+        </dependency>
+        <dependency>
+            <groupId>io.grpc</groupId>
+            <artifactId>grpc-stub</artifactId>
+            <version>1.45.0</version>
         </dependency>
     </dependencies>
