+++ pom.xml
@@ -17,6 +17,7 @@
         <java.version>14</java.version>
         <maven.compiler.source>14</maven.compiler.source>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <javalib.version>0.0.6</javalib.version>
     </properties>
 
     <modules>
@@ -58,11 +59,12 @@
             <artifactId>validation-api</artifactId>
         </dependency>
         <!-- https://mvnrepository.com/artifact/com.alibaba/fastjson -->
-        <dependency>
+    <!--    <dependency>
             <groupId>com.alibaba</groupId>
             <artifactId>fastjson</artifactId>
             <version>1.2.83</version>
-        </dependency>
+        </dependency>-->
+
     </dependencies>
     <build>
         <plugins>
